+++ src/main/java/by/itechart/javalab/dao/ContactModificationDao.java
@@ -10,7 +10,7 @@ import java.util.List;
 public interface ContactModificationDao {
     Contact addNewContact(Contact contact) throws DaoException;
     Contact updateContact(Contact contact) throws DaoException;
-    void deleteContacts(Integer contactsId[]) throws DaoException;
+    void deleteContacts(Long contactsId[]) throws DaoException;
     void saveContactPhones(List<ContactPhone> phones) throws DaoException;
     void deleteContactPhones(List<ContactPhone> phones) throws DaoException;
     List<ContactAttachment> saveContactAttachments(List<ContactAttachment> attachments) throws DaoException;
